+++ src/modules/all.less
@@ -15,3 +15,6 @@
 @import "./esl-scrollable-tabs/core.less";
 
 @import "./esl-alert/core.less";
+
+@import "./esl-select/core.less";
+@import "./esl-select-list/core.less";
