+++ src/modules/all.ts
@@ -20,3 +20,6 @@ export * from './esl-scrollbar/core';
 
 export * from './esl-alert/core';
 
+export * from './esl-select/core';
+export * from './esl-select-list/core';
+
