+++ src/modules/draft/all.less
@@ -1,4 +1,1 @@
-@import "./esl-select/core.less";
-@import "./esl-select-list/core.less";
-
 @import "./esl-carousel/all.less";
