+++ src/modules/esl-select-list/core/esl-select-list.ts
@@ -1,10 +1,11 @@
-import {attr, boolAttr} from '../../../esl-base-element/core';
-import {bind} from '../../../esl-utils/decorators/bind';
-import {ESLScrollbar} from '../../../esl-scrollbar/core';
+import {attr, boolAttr} from '../../esl-base-element/core';
+import {bind} from '../../esl-utils/decorators/bind';
+import {ExportNs} from '../../esl-utils/environment/export-ns';
+import {ESLScrollbar} from '../../esl-scrollbar/core';
 import {ESLSelectItem} from './esl-select-item';
-
 import {ESLSelectWrapper} from './esl-select-wrapper';
 
+@ExportNs('SelectList')
 export class ESLSelectList extends ESLSelectWrapper {
   public static readonly is = 'esl-select-list';
   public static get observedAttributes() {
