+++ src/modules/esl-select/core/esl-select-renderer.less
